LPD-86259 Merge MissingReferenceManager into EmptyModelManager
The single-method MissingReferenceManager added earlier on this branch overlapped in purpose with the existing EmptyModelManager (both record export/import findings against the report entry service during an import). Folding reportMissingReference into EmptyModelManager removes the parallel OSGi service, its API package, its bnd Export-Package entry, and the Snapshot indirection LogUtil needed to dereference it.

The new reportMissingReference method may look out of place on a class named EmptyModelManager, but I prefer to keep both responsibilities together because it simplifies the wiring; the class is effectively a manager for reporting missing or empty references during an import, and a future rename can make that scope explicit if needed.

The impl reuses the GroupLocalService and ClassNameLocalService references already present on EmptyModelManagerImpl. companyId is now derived from the resolved Group (consistent with the empty-model path) rather than from the ExportImportConfiguration thread local, which also lets the method short-circuit cleanly when groupId is 0 or otherwise unknown.

The empty.model packageinfo bumps to 2.2.0 (additive method on @ProviderType interface); the report-api report.missing.reference package is gone entirely.
