--- modules/apps/export-import/export-import-report-api/src/main/java/com/liferay/exportimport/report/missing/reference/MissingReferenceManager.java
@@ -1,16 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.exportimport.report.missing.reference;
-
-/**
- * @author Petteri Karttunen
- */
-public interface MissingReferenceManager {
-
-	public void reportMissingReference(
-		String className, String externalReferenceCode, long groupId);
-
-}
