--- modules/apps/export-import/export-import-report-api/src/main/resources/com/liferay/exportimport/report/missing/reference/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
