+++ modules/apps/headless/headless-admin-site/headless-admin-site-impl/build.gradle
@@ -17,7 +17,6 @@ dependencies {
 	compileOnly project(":apps:client-extension:client-extension-type-api")
 	compileOnly project(":apps:document-library:document-library-api")
 	compileOnly project(":apps:export-import:export-import-api")
-	compileOnly project(":apps:export-import:export-import-report-api")
 	compileOnly project(":apps:fragment:fragment-api")
 	compileOnly project(":apps:fragment:fragment-collection-filter-api")
 	compileOnly project(":apps:fragment:fragment-entry-processor:fragment-entry-processor-api")
