+++ modules/apps/headless/headless-admin-site/headless-admin-site-impl/src/main/java/com/liferay/headless/admin/site/internal/util/LogUtil.java
@@ -5,12 +5,11 @@
 
 package com.liferay.headless.admin.site.internal.util;
 
-import com.liferay.exportimport.report.missing.reference.MissingReferenceManager;
+import com.liferay.exportimport.kernel.empty.model.EmptyModelManagerUtil;
 import com.liferay.headless.admin.site.dto.v1_0.ItemExternalReference;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.module.service.Snapshot;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.vulcan.scope.Scope;
 
@@ -35,7 +34,7 @@ public class LogUtil {
 			_log.warn(sb.toString());
 		}
 
-		_reportMissingReference(
+		EmptyModelManagerUtil.reportMissingReference(
 			modelClass.getName(), modelExternalReferenceCode, scopeId);
 	}
 
@@ -58,7 +57,7 @@ public class LogUtil {
 			_log.warn(sb.toString());
 		}
 
-		_reportMissingReference(className, null, 0);
+		EmptyModelManagerUtil.reportMissingReference(className, null, 0);
 	}
 
 	public static void logOptionalReference(
@@ -88,27 +87,10 @@ public class LogUtil {
 			_log.warn(sb.toString());
 		}
 
-		_reportMissingReference(className, externalReferenceCode, scopeId);
-	}
-
-	private static void _reportMissingReference(
-		String className, String externalReferenceCode, long scopeId) {
-
-		MissingReferenceManager missingReferenceManager =
-			_missingReferenceManagerSnapshot.get();
-
-		if (missingReferenceManager == null) {
-			return;
-		}
-
-		missingReferenceManager.reportMissingReference(
+		EmptyModelManagerUtil.reportMissingReference(
 			className, externalReferenceCode, scopeId);
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(LogUtil.class);
-
-	private static final Snapshot<MissingReferenceManager>
-		_missingReferenceManagerSnapshot = new Snapshot<>(
-			LogUtil.class, MissingReferenceManager.class);
 
 }
