+++ portal-kernel/src/com/liferay/exportimport/kernel/empty/model/packageinfo
@@ -1,1 +1,1 @@
-version 2.1.0
+version 2.2.0
